+++ src/components/organisms/table/table.style.ts
