+++ src/components/templates/login/LoginTitle.tsx
@@ -1,0 +1,8 @@
+const LoginTitle = () => (
+  <div className='flex flex-col gap-[8px] pb-[30px]'>
+    <p className='text-[14px]'>오늘은 회사에서 어떤 일들이 펼쳐질까</p>
+    <span className='text-[34px] font-[700]'>오늘의 워크로그</span>
+  </div>
+);
+
+export default LoginTitle;
