+++ src/stories/components/Button.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryFn } from '@storybook/react';
 
-import { Button } from '~/components/button/Button';
+import { Button } from '~/components/molecules/button/Button';
 
 const meta: Meta<typeof Button> = {
   title: 'Components/Button',
