+++ src/stories/pages/Login.stories.tsx
@@ -1,14 +1,14 @@
 import { Meta, StoryFn } from '@storybook/react';
 
-import Login from '../../pages/login';
+import LoginPage from '../../pages/login';
 
-import { NonAuthLayout } from '~/components/layout/NonAuthLayout';
+import { NonAuthLayout } from '~/components/templates/layout/NonAuthLayout';
 
 import { loginHandler } from '~/tests/handlers/auth';
 
-const meta: Meta<typeof Login> = {
+const meta: Meta<typeof LoginPage> = {
   title: 'Pages/Login',
-  component: Login,
+  component: LoginPage,
   decorators: [
     (Story) => (
       <NonAuthLayout>
@@ -25,4 +25,4 @@ const meta: Meta<typeof Login> = {
 
 export default meta;
 
-export const Default: StoryFn<typeof Login> = () => <Login />;
+export const Default: StoryFn<typeof LoginPage> = () => <LoginPage />;
